requirements.txt
cryptography==44.0.1
click==8.1.7
python-dotenv~=1.1.1